Add portfolio index.html
Initial commit for my personal portfolio website.

This project is created and maintained by Awais Iftikhar, a Computer Science student,
Web Developer, Graphic Designer and Cold Calling expert. The goal of this repository is to showcase my
skills, highlight my projects, and provide a way for people to connect with me.

Features added in this commit:
- Created index.html as the main landing page
- Added header section with my name and profession
- Added placeholder for "Projects" section
- Added footer with contact information

This portfolio will continue to evolve as I add more projects and refine the design.