index.html
Initial commit for my personal portfolio website.

This project is created and maintained by Awais Iftikhar, a Computer Science student,
Web Developer, and Graphic Designer. The goal of this repository is to showcase my
skills, highlight my projects, and provide a way for people to connect with me.
This portfolio will continue to evolve as I add more projects and refine the design.